+++ airbyte-cron/src/main/java/io/airbyte/cron/jobs/WorkspaceCleaner.java
@@ -8,14 +8,13 @@ import static io.airbyte.cron.MicronautCronRunner.SCHEDULED_TRACE_OPERATION_NAME
 
 import datadog.trace.api.Trace;
 import io.airbyte.commons.envvar.EnvVar;
-import io.airbyte.config.Configs;
-import io.airbyte.config.EnvConfigs;
 import io.airbyte.metrics.lib.ApmTraceUtils;
 import io.airbyte.metrics.lib.MetricAttribute;
 import io.airbyte.metrics.lib.MetricClient;
 import io.airbyte.metrics.lib.MetricTags;
 import io.airbyte.metrics.lib.OssMetricsRegistry;
 import io.micronaut.context.annotation.Requires;
+import io.micronaut.context.annotation.Value;
 import io.micronaut.context.env.Environment;
 import io.micronaut.scheduling.annotation.Scheduled;
 import jakarta.inject.Singleton;
@@ -28,34 +27,34 @@ import java.time.OffsetDateTime;
 import java.util.Date;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
-import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.filefilter.AgeFileFilter;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * Delete old files that accumulate in docker.
  */
 @Singleton
-@Slf4j
 @Requires(notEnv = Environment.KUBERNETES)
 public class WorkspaceCleaner {
 
-  private final Path workspaceRoot;
+  private static final Logger log = LoggerFactory.getLogger(WorkspaceCleaner.class);
+
+  private final Path workspaceRootPath;
   private final long maxAgeFilesInDays;
   private final MetricClient metricClient;
 
   public static final String DEFAULT_TEMPORAL_HISTORY_RETENTION_IN_DAYS = "30";
 
-  WorkspaceCleaner(final MetricClient metricClient) {
-    log.info("Creating workspace cleaner");
-
-    // TODO Configs should get injected through micronaut
-    final Configs configs = new EnvConfigs();
-
-    this.workspaceRoot = configs.getWorkspaceRoot();
+  public WorkspaceCleaner(
+                          final MetricClient metricClient,
+                          @Value("${airbyte.workspace.root}") final String workspaceRoot) {
+    this.workspaceRootPath = Path.of(workspaceRoot);
     // We align max file age on temporal for history consistency
     // It might make sense configure this independently in the future
-    this.maxAgeFilesInDays = Integer.parseInt(EnvVar.TEMPORAL_HISTORY_RETENTION_IN_DAYS.fetch(DEFAULT_TEMPORAL_HISTORY_RETENTION_IN_DAYS));
+    this.maxAgeFilesInDays = Integer.parseInt(
+        EnvVar.TEMPORAL_HISTORY_RETENTION_IN_DAYS.fetch(DEFAULT_TEMPORAL_HISTORY_RETENTION_IN_DAYS));
     this.metricClient = metricClient;
   }
 
@@ -70,24 +69,29 @@ public class WorkspaceCleaner {
   public void deleteOldFiles() throws IOException {
     final Date oldestAllowed = getDateFromDaysAgo(maxAgeFilesInDays);
     log.info("Deleting files older than {} days ({})", maxAgeFilesInDays, oldestAllowed);
-    metricClient.count(OssMetricsRegistry.CRON_JOB_RUN_BY_CRON_TYPE, 1, new MetricAttribute(MetricTags.CRON_TYPE, "workspace_cleaner"));
+    metricClient.count(OssMetricsRegistry.CRON_JOB_RUN_BY_CRON_TYPE, 1,
+        new MetricAttribute(MetricTags.CRON_TYPE, "workspace_cleaner"));
 
     ApmTraceUtils.addTagsToTrace(Map.of("oldest_date_allowed", oldestAllowed, "max_age", maxAgeFilesInDays));
 
     final AtomicInteger counter = new AtomicInteger(0);
-    Files.walk(workspaceRoot)
-        .map(Path::toFile)
-        .filter(f -> new AgeFileFilter(oldestAllowed).accept(f))
-        .forEach(file -> {
-          log.debug("Deleting file: " + file.toString());
-          FileUtils.deleteQuietly(file);
-          counter.incrementAndGet();
-          final File parentDir = file.getParentFile();
-          if (parentDir.isDirectory() && parentDir.listFiles().length == 0) {
-            FileUtils.deleteQuietly(parentDir);
-          }
-        });
-    log.info("deleted {} files", counter.get());
+
+    // Check if workspace root path exists. It may not exist during tests.
+    if (workspaceRootPath.toFile().exists()) {
+      Files.walk(workspaceRootPath)
+          .map(Path::toFile)
+          .filter(f -> new AgeFileFilter(oldestAllowed).accept(f))
+          .forEach(file -> {
+            log.debug("Deleting file: {}", file);
+            FileUtils.deleteQuietly(file);
+            counter.incrementAndGet();
+            final File parentDir = file.getParentFile();
+            if (parentDir.isDirectory() && parentDir.listFiles().length == 0) {
+              FileUtils.deleteQuietly(parentDir);
+            }
+          });
+      log.info("deleted {} files", counter.get());
+    }
   }
 
   private static Date getDateFromDaysAgo(final long daysAgo) {
